+++ lib/meta/types.dart
@@ -14,8 +14,13 @@ class AttributeType extends ConceptEntity<AttributeType> {
   AttributeType(this.domain, String typeCode) {
     if (typeCode == 'Email') {
       base = 'String';
+      length = 48;
+    } else if (typeCode == 'Uri') {
+      base = 'Uri';
+      length = 80;
     } else if (typeCode == 'Other') {
       base = 'dynamic';
+      length = 64;
     } else {
       base = typeCode;
     }